+++ CHANGELOG.md
@@ -4,7 +4,7 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
-## [Unreleased]
+## [1.14.1] - 2026-03-29
 ### Fixed
  - Fix Steam installation being prompted twice when using `protontricks-launch`
  - Fix Steam Runtime crash if it's installed under a blacklisted root directory (eg. `/run/<path>` instead of `/run/media/<path>`)
@@ -343,7 +343,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 ### Added
  - The last release of Protontricks maintained by [@Sirmentio](https://github.com/Sirmentio).
 
-[Unreleased]: https://github.com/Matoking/protontricks/compare/1.14.0...HEAD
+[Unreleased]: https://github.com/Matoking/protontricks/compare/1.14.1...HEAD
+[1.14.1]: https://github.com/Matoking/protontricks/compare/1.14.0...1.14.1
 [1.14.0]: https://github.com/Matoking/protontricks/compare/1.13.1...1.14.0
 [1.13.1]: https://github.com/Matoking/protontricks/compare/1.13.0...1.13.1
 [1.13.0]: https://github.com/Matoking/protontricks/compare/1.12.1...1.13.0